feature: soaking enable-attribute for Beaker XMLs
The idea of this patch is this: test is either finished onboarding and
is being added to resulting Beaker XML normally, or it has
enabled|disabled soaking.

If the test soaking is enabled, skt gathers aggregate results of that
test's runcount into redis.
If the test soaking is disabled, we are not gathering any data, but we
know we may want to start (again).

Normal (onboarded tests) should not be affected in any way.

This patch modifies test_targeted and other files to insert

soaking="{val}"

into those test <task /> elements that are enabled for soaking.

A test that has soaking enabled or disabled (isn't onboarded yet) has:

    "soaking": 1
or

    "soaking": 0

If the test doesn't have that, then it's a test that was already
onboarded.
A test that was already onboarded shouldn't be moved to "soaking",
unless it has been failing in some serious way and its maintainer has
failed to care enough to update it.

V2:
* flake8 and pylint fixes
* some superficial test coverage
V3:
* more trivial fixes
V4:
* remove runcount from kpet-db

Signed-off-by: Jakub Racek <jracek@redhat.com>