Switch to using data.Run class in run command
Move most of the "run" command logic to a couple classes in the data
module: TestSuiteRun and Run.

The TestSuiteRun class represents an execution of a test suite, and its
instances are to be exposed directly to templates later.

The Run class represents an execution of all the tests in the database,
and allows listing the resulting cases and rendering the Beaker XML
which would execute the run. It will not be exposed to templates itself.

Exporting lists test case templates in data.Run.generate is to be
removed once the test suites are exposed to the templates and they make
use of them.