Introduce run module
Introduce a "run" module - a collection of classes representing an
execution of tests from an instance of a database from the "data"
module. Essentially, the data that is made available to templates after
all the database is massaged according to execution logic.

Previously the idea was to expose the database as is to the tests and
supply them with helper functions for processing it, with the intention
of keeping it obvious where the data comes from (i.e. YAML files) and in
which shape (same shape as in YAML files), but it's useless if all the
logic is hiding in Python code anyway, and implementing all that logic
in Jinja would be too much.

The current module is simply copying the existing logic, but it will be
expanded in further commits.