data: Introduce suite/case list matching
Introduce functions matching suites and cases against a source path set,
but returning a list instead of a set. This would allow us to preserve
the database order of suites and cases. Keep the set-returning functions
for now, to be removed when they're no longer needed by the database's
templates.