Add support for excluding cases from baseline runs
Add support for specifying a "specific" flag at the global, a suite, and
a case level. The most-specific specified "specific" flag (ha!) wins.
If none are specified, it is assumed to be false. If it's true, then the
case doesn't match an empty set of source file paths. That is, it won't
get executed on baseline runs (runs which have no patches specified).