Implement generic suite/case matching
Implement generic suite/case matching. For the start it supports
matching suites/cases against a set of source files, whether they're
specific or wildcard (all files, for baseline testing), against an
architecture, and against a tree.

In fact, all of these support matching against a set internally, just
for the simplicity of the implementation. However neither the
command-line interface nor the template interface expose that. Although
the architecture set matching might come in useful later if we decide to
generate a single job XML for all architectures.

This also replaces the "specific" field used in cases/suites and
database previously, without a data upgrade path, since it wasn't used
yet.