schema: Implement Reduction schema
Implement the "Reduction" schema, a derivative of "Choice" and identical
to "Succession" for the case of one or two schemas, but different for
three and above. It allows specifying a bunch of alternative schemas and
one general schema which the former could be converted to. Could be used
for describing nodes accepting multiple alternative types, but presented
as a single type to the code.