adjust all tests for suites tasks
Integration and unit tests for kpet have a lot of assets. Currently, to take
advantage of the pending kpet-db MR that gets rid of all task XML
snippets, 'maintainer: ' has to be added to the yamls. Field fetch_suffix
is not mandatory, but desirable.
Make maintainers non-empty optional list of strings.

The jinja template is adjusted to generate XML snippets instead of
loading them from files.

Changes over previous versions:
* overhaul completely
* pylint fixes
* add fixes requested in review
* changes should pass tests independently

Signed-off-by: Jakub Racek <jracek@redhat.com>