allow specifying url_suffix per-case
Currently, if url_suffix is set in suite, all cases get tests-beaker
fetch element, with url_suffix set. If url_suffix is missing, the fetch
element is not present.

This patch allows to set url_suffix per-case. If it is present in both
suite and case, then case overrides suite.

This requires complementary kpet-db change.

Signed-off-by: Jakub Racek <jracek@redhat.com>