add hostname attribute for overwriting hostRequires
This is kpet change necessary to implement FASTMOVING-740.

This patch adds hostname attribute into host/suite/case. This attribute
is the hostname to use for this host/suite/case.

This allows kpet-db jinja template to insert <hostRequires force="..." />
element insteadof any normal hostRequires.
This change is independent from kpet-db, but in order to work, kpet-db
needs tiny either/or change in template

Signed-off-by: Jakub Racek <jracek@redhat.com>