Simplify tox configuration
Instead of having different tox environments for the different
linting tools we're using, have only two environments: one to run
linting + tests, which will be the default, and one to run the
coverage-statistics gathering command (only used in CI).

Also, stop forcing Python to be 3.7 in the tox configuration, as it's
not necessary (and would break tests and such in any developer machine
that doesn't have Python 3.7 installed) and we're forcing the version
on Travis anyway.

Another positive outcome is that the Travis configuration is much
simplified, so it's easier to replicate in other CI systems if needed,
or simply understand how it works and how to change it.

While at it, clean up setup.py a bit.