Accept "name" as an optional property in suites
We are going to add a name to test suites. For that to happen,
however, we first need to _allow_ it without requiring it, then change
kpet-db to add it, then change the property to be required. This is
the first step.

As part of this, expose the test suite name to Beaker XML templates so
that it can be exported in the kpet-db templates.

Ref. FASTMOVING-1278