run: Support specifying template variables
Add support for specifying extra variables for templates on "kpet run
generate" command line. Add a separate sub-command "kpet variable" to
allow listing available variables and their defaults, which are
specified in the database. Verify no unknown variables are specified,
and all required variables are specified. Use defaults where set.

This is to replace the use of TESTS_BEAKER_ZIP_URL environment variable
and getenv, as well as the -g/--group option to "kpet run generate".