Clean up validation error messages
Revamp the way we report validation exceptions. In validation code we
often throw exceptions inside exception handlers, and because of the
way we were reporting them, the error messages the user saw had lots
of repeated bits, to the point of making the output unreadable (and
maybe even making the user think that there's a bug in
kpet!). Instead, show a more helpful error message for the user which
doesn't repeat information.

A couple of tests had to be tweaked slightly to adapt to the
differences in error reporting.

Ref. FASTMOVING-1456