Add a kpet.run class for test cases
Other "data" classes have a simplified version in kpet.run. This is
the version that is passed to the templates, and only contain
attributes that are safe to put in templates. The exception for that
was Case, so create a kpet.run.Case class to make sure we only expose
the things we want. Also pass it to the kpet.run.Suite constructor,
which now excepts a list of those objects.

Ref. FASTMOVING-1443