Make sure there is a maintainer for every testcase
Now that we can specify maintainers at the test case level, we want to
allow specifying _only_ at that level and not at the suite level. That
failed before, so fix that and make sure that every test case has a
maintainer, either from the case itself, or from its containing suite.