Rename hostRequires_list to host_requires_list
Although it does hold a list of "hostRequires", it's a bit ugly that
we end up with a variable that has both mixedCase and snake_case in
the same name. So, default to the normal Python convention and use
"host_requires_list" for the variable.

However, need to keep hostRequires_list temporarily for compatibility
reasons.

Ref. FASTMOVING-1612